+++ include/lepto/signal.hpp
@@ -45,12 +45,11 @@
 
 #else
 
-#if 0
-   // By default, use virtual signals supportunǵ fungtions and methods
-   #if ! defined ( CONFIG_LEPTO_SIGNAL_FUNCTION ) && ! defined ( CONFIG_LEPTO_SIGNAL_METHOD )
-      #define CONFIG_LEPTO_SIGNAL_FUNCTION      1
-      #define CONFIG_LEPTO_SIGNAL_METHOD        1
-   #endif
+// By default, use virtual signals supportunǵ methods but not functions. This
+// may lead to CFunctorMethodAsFunction being used. 
+#if ! defined ( CONFIG_LEPTO_SIGNAL_FUNCTION ) && ! defined ( CONFIG_LEPTO_SIGNAL_METHOD )
+   #undef CONFIG_LEPTO_SIGNAL_FUNCTION
+   #define CONFIG_LEPTO_SIGNAL_METHOD        1
 #endif
 
 #if IS_ENABLED( CONFIG_LEPTO_SIGNAL_FUNCTION ) && IS_ENABLED( CONFIG_LEPTO_SIGNAL_METHOD )
@@ -60,6 +59,10 @@
    // Dont use virtual emit methods
    #define LEPTO_SIGNAL_VIRTUAL
    #define LEPTO_SIGNAL_DO_VIRTUAL           0
+#endif
+
+#if LEPTO_SIGNAL_DO_VIRTUAL || IS_ENABLED( CONFIG_LEPTO_SIGNAL_FUNCTION ) || IS_ENABLED( CONFIG_LEPTO_SIGNAL_CHAIN )
+   #define LEPTO_SIGNAL_FUNCTOR_ALLOCATED    1
 #endif
 
 // The template 'sclotClass' can never be part of the signals class declaration
@@ -119,45 +122,71 @@ class CFunctorMethod final
       #endif
 {
    public:
-      slotClass *slotObject;
-      sigReturn (slotClass::*methodPtr)( sigTypes ... args );
+      slotClass *m_slotObject;
+      sigReturn (slotClass::*m_methodPtr)( sigTypes ... args );
 
-      constexpr CFunctorMethod( slotClass *_slotObject, sigReturn (slotClass::*_methodPtr)( sigTypes ... args ))
-         :slotObject(_slotObject)
-         ,methodPtr(_methodPtr)
+      constexpr CFunctorMethod( slotClass *slotObject, sigReturn (slotClass::*methodPtr)( sigTypes ... args ))
+         :m_slotObject( slotObject )
+         ,m_methodPtr( methodPtr )
       {
       }
 
       LEPTO_SIGNAL_VIRTUAL
       sigReturn emitSignal( sigTypes ... args ) const //final
       {
-         return( (slotObject->*(this->methodPtr))( args... ) );
+         return( (m_slotObject->*(this->m_methodPtr))( args... ) );
       }
 };
 
 template <typename sigReturn, typename ... sigTypes>
 class CFunctorMethodAsFunction final
+      #if LEPTO_SIGNAL_DO_VIRTUAL
+          : public CFunctor<sigReturn, sigTypes...>
+      #endif
 {
+      friend class CSignal<sigReturn, sigTypes...>;
+      
+      #if IS_ENABLED( CONFIG_LEPTO_SIGNAL_CHAIN )
+         CFunctorMethodAsFunction *m_next=nullptr;
+      #endif
+
    public:
-   void* slotObject;
-   sigReturn (*methodPtr)( void *, sigTypes ... args );
+   void* m_slotObject;
+   sigReturn (*m_methodPtr)( void *, sigTypes ... args );
    
    #pragma GCC diagnostic push
    #pragma GCC diagnostic ignored "-Wpmf-conversions"
    
    template <class slotClass>
-   constexpr CFunctorMethodAsFunction( slotClass* _slotObject, sigReturn (slotClass::*_methodPtr)( sigTypes ... args ))
-       :slotObject(_slotObject)
-       ,methodPtr((sigReturn (*)( void *, sigTypes ... args ))_methodPtr)
+   constexpr CFunctorMethodAsFunction( slotClass* slotObject, sigReturn (slotClass::*methodPtr)( sigTypes ... args ))
+       :m_slotObject( slotObject )
+       ,m_methodPtr( (sigReturn (*)( void *, sigTypes ... args ))methodPtr )
    {
    }
+
+   #if 1
+   constexpr CFunctorMethodAsFunction( )
+       :m_methodPtr( nullptr )
+   {
+   }
+   template <class slotClass>
+   void connect( slotClass* _slotObject, sigReturn (slotClass::*_methodPtr)( sigTypes ... args ))
+   {
+      m_slotObject=_slotObject;
+      m_methodPtr=(sigReturn (*)( void *, sigTypes ... args ))_methodPtr;
+   }
+   #endif
    
    #pragma GCC diagnostic pop
    
    LEPTO_SIGNAL_VIRTUAL
        sigReturn emitSignal( sigTypes ... args ) const //final
    {
-      return( (*(this->methodPtr))( slotObject, args... ) );
+      return( (*(this->m_methodPtr))( m_slotObject, args... ) );
+   }
+   const bool isConnected() const
+   {
+      return( m_methodPtr != nullptr );
    }
 };
 
@@ -204,6 +233,16 @@ class CSignalMethod
 #endif // 0
 
 
+#if LEPTO_SIGNAL_DO_VIRTUAL
+   #define CFunctorAbstract CFunctor
+#elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_FUNCTION )
+   #define CFunctorAbstract CFunctorFunction
+#elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_METHOD )
+   #define CFunctorAbstract CFunctorMethodAsFunction
+   #define CFunctorMethodConcrete CFunctorMethodAsFunction
+#endif
+
+
 // The template 'sclotClass' can never be part of the signals class declaration
 // because the class of the slot is not known at this timepoint of course.
 template <typename sigReturn, typename ... sigTypes>
@@ -218,7 +257,11 @@ class CSignal
          const CFunctorFunction<sigReturn, sigTypes...> *m_pFunctor;
       #elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_METHOD )
          //const CFunctorMethod<sigReturn, slotClass, sigTypes...>* m_pFunctor;
-         const CFunctorMethodAsFunction<sigReturn, sigTypes...> *m_pFunctor;
+         #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+            CFunctorMethodAsFunction<sigReturn, sigTypes...> *m_pFunctor;
+         #else
+            CFunctorMethodAsFunction<sigReturn, sigTypes...> m_pFunctor;
+         #endif
       #else
          #error "Could not check signal configuration"
       #endif
@@ -226,7 +269,10 @@ class CSignal
    public:
 
       constexpr CSignal()
-         :m_pFunctor{ nullptr }
+         #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+            :m_pFunctor{ nullptr }
+         #else
+         #endif
       {
       };
       ~CSignal()
@@ -245,19 +291,37 @@ class CSignal
       #if LEPTO_SIGNAL_DO_VIRTUAL || 1
       template <class slotClass >
       #endif
-      void connect( slotClass *slotObject, sigReturn (slotClass::*_methodPtr)( sigTypes ... args ))
+      void connect( slotClass *slotObject, sigReturn (slotClass::*methodPtr)( sigTypes ... args ))
       {
          #if IS_ENABLED( CONFIG_LEPTO_SIGNAL_CHAIN )
-         CFunctor<sigReturn, sigTypes...> **pFunctor=&m_pFunctor;
-         while(*pFunctor)
-         {
-            pFunctor=&((*pFunctor)->m_next);
-         }
-         *pFunctor=new CFunctorMethod<sigReturn, slotClass, sigTypes...>(slotObject, _methodPtr);
-         #else
-         lAssert( m_pFunctor == nullptr );
-         m_pFunctor=new CFunctorMethodAsFunction( slotObject, _methodPtr );
          
+            #if LEPTO_SIGNAL_DO_VIRTUAL
+               CFunctor<sigReturn, sigTypes...> **pFunctor=&m_pFunctor;
+            #elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_FUNCTION )
+               CFunctorFunction<sigReturn, sigTypes...>**pFunctor=&m_pFunctor;
+            #elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_METHOD )
+               CFunctorMethodAsFunction<sigReturn, sigTypes...>**pFunctor=&m_pFunctor;
+            #else
+               #error "Could not check signal configuration"
+            #endif
+            
+            while(*pFunctor)
+            {
+               pFunctor=&((*pFunctor)->m_next);
+            }
+            *pFunctor=new CFunctorMethodAsFunction(slotObject, methodPtr);
+         #else
+            #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+               lAssert( m_pFunctor == nullptr );
+            #else
+               lAssert( ! m_pFunctor.isConnected() );
+            #endif
+               
+            #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+               m_pFunctor=new CFunctorMethodAsFunction( slotObject, methodPtr );
+            #else
+               m_pFunctor.connect(slotObject, methodPtr);
+            #endif
          #endif
       };
       #endif
@@ -273,17 +337,37 @@ class CSignal
       void emitSignal( sigTypes ... args ) const
       {
          #if IS_ENABLED( CONFIG_LEPTO_SIGNAL_CHAIN )
-         const CFunctor<sigReturn, sigTypes...> * const*pFunctor=&m_pFunctor;
-         while(*pFunctor)
-         {
-            (*pFunctor)->emitSignal( args ... );
-            pFunctor=&((*pFunctor)->m_next);
-         }
+            #if ! LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+               #error CONFIG_LEPTO_SIGNAL_CHAIN Only works with allocated functors
+            #endif
+         
+            #if LEPTO_SIGNAL_DO_VIRTUAL
+               const CFunctor<sigReturn, sigTypes...> * const*pFunctor=&m_pFunctor;
+            #elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_FUNCTION )
+               const CFunctorFunction<sigReturn, sigTypes...>* const *pFunctor=&m_pFunctor;
+            #elif IS_ENABLED( CONFIG_LEPTO_SIGNAL_METHOD )
+               const CFunctorMethodAsFunction<sigReturn, sigTypes...>* const *pFunctor=&m_pFunctor;
+            #else
+               #error "Could not check signal configuration"
+            #endif
+
+            while(*pFunctor)
+            {
+               (*pFunctor)->emitSignal( args ... );
+               pFunctor=&((*pFunctor)->m_next);
+            }
          #else
-         if( m_pFunctor )
-         {
-            m_pFunctor->emitSignal( args ... );
-         }
+            #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+               if( m_pFunctor )
+               {
+                  m_pFunctor->emitSignal( args ... );
+               }
+            #else
+               if( m_pFunctor.isConnected() )
+               {
+                  m_pFunctor.emitSignal( args ... );
+               }
+            #endif
          #endif
          
          return;
@@ -320,10 +404,17 @@ class CSignal
        */
       sigReturn emitSingle( sigTypes ... args ) const
       {
-         if( m_pFunctor )
-         {
-            return( m_pFunctor->emitSignal( args ... ) );
-         }
+         #if LEPTO_SIGNAL_FUNCTOR_ALLOCATED
+            if( m_pFunctor )
+            {
+               return( m_pFunctor->emitSignal( args ... ) );
+            }
+         #else
+            if( m_pFunctor.isConnected() )
+            {
+               return( m_pFunctor.emitSignal( args ... ) );
+            }
+         #endif
          return(0);
       }